refactor encounter service naming, add save catch method
